move type aliases
